Avatar refactor * exports
